fix(executor): Schema-aware WHERE alias resolution and USING clause support
- Add resolve_where_aliases_with_schema() for SQLite-compatible alias handling
  Table column names now correctly take precedence over SELECT aliases
  Example: SELECT COUNT(*) AS col1 FROM t WHERE col1 > 0
  col1 in WHERE refers to table column, not the alias

- Add USING clause column deduplication for JOIN operations
  USING clause joins now properly show join columns only once (from left side)

- Export order module as pub(crate) for cross-module access

- Update aggregate function handling and tests for consistency

- Update benchmark data

+++ crates/vibesql-executor/src/tests/join_aggregation.rs
@@ -332,13 +332,13 @@ fn test_left_join_with_group_by_avg_salary() {
         .unwrap();
 
     // Engineering: (95000 + 87000) / 2 = 91000
-    assert_eq!(engineering_row.get(1).unwrap(), &SqlValue::Numeric(91000.0));
+    assert_eq!(engineering_row.get(1).unwrap(), &SqlValue::Double(91000.0));
 
     // Sales: (75000 + 72000) / 2 = 73500
-    assert_eq!(sales_row.get(1).unwrap(), &SqlValue::Numeric(73500.0));
+    assert_eq!(sales_row.get(1).unwrap(), &SqlValue::Double(73500.0));
 
     // HR: 65000 / 1 = 65000
-    assert_eq!(hr_row.get(1).unwrap(), &SqlValue::Numeric(65000.0));
+    assert_eq!(hr_row.get(1).unwrap(), &SqlValue::Double(65000.0));
 }
 
 #[test]
